backend/requirements.txt
dedalus-labs>=0.1.0a9
fastapi[standard]>=0.121.0
python-dotenv>=1.2.1
httpx>=0.28.1
openai>=1.0.0
snowflake-connector-python>=3.7.0

